+++ src/olorenchemengine/dataset.py
@@ -392,4 +392,24 @@ class BaseKFold(BaseDatasetTransform):
     def transform(self, dataset: BaseDataset, random_state: int = 42, *args, **kwargs):
         """Splits document into folds, identified by 1, ..., n_splits in the 'cv' column."""
         pass
+
+class RandomKFold(BaseDatasetTransform):
     
+    def transform(self, dataset: BaseDataset, *args, random_state: int = 42, **kwargs):
+        np.random.seed(random_state)
+        fold_num = np.repeat(1 + np.arange(self.n_splits), 1 + len(dataset.data) // self.n_splits)[:len(dataset.data)]
+        dataset.data["cv"] = np.random.permutation(fold_num)
+        return dataset
+
+from rdkit.Chem.Scaffolds import MurckoScaffold
+
+class ScaffoldKFold(BaseDatasetTransform):
+    
+    def transform(self, dataset: BaseDataset, *args, random_state: int = 42, **kwargs):
+        np.random.seed(random_state)
+        scaffolds = [MurckoScaffold.GetScaffoldForMol(Chem.MolFromSmiles(x)) for x in dataset.data[dataset.structure_col]]
+        folds = np.random.permutation(self.n_splits * np.ones(len(scaffolds), dtype=int))
+        for fold, scaffold in enumerate(sorted(set(scaffolds))):
+            folds[np.array(scaffolds) == scaffold] = 1 + fold % self.n_splits
+        dataset.data["cv"] = folds
+        return dataset
